Portable registration database
