+++ lib/database.js
@@ -3,6 +3,8 @@ import { createHash, randomBytes, createCipheriv, createDecipheriv } from 'crypt
 import Listener from 'pg-listen';
 import logger from './logger.js';
 import { getHandler } from './handlers/index.js';
+import { initPhoneRegistration } from './models/phone-registration.js';
+import { encryptSecret, decryptSecret } from './utils/credentials.js';
 
 const schemaVersion = 17;
 
@@ -65,51 +67,6 @@ let listener = new Listener(LISTENER_CONNECTION);
 
 let streamIds = {};
 
-// Simple AES-256-GCM helpers for encrypting credentials at rest
-const deriveKey = (secret) => secret && createHash('sha256').update(String(secret)).digest();
-const encryptionKey = deriveKey(CREDENTIALS_KEY);
-
-function encryptSecret(plainText) {
-  if (plainText == null) return null;
-  try {
-    if (!encryptionKey) {
-      logger.warn('CREDENTIALS_KEY not set; storing password in plaintext');
-      return plainText;
-    }
-    const iv = randomBytes(12);
-    const cipher = createCipheriv('aes-256-gcm', encryptionKey, iv);
-    const ciphertext = Buffer.concat([cipher.update(String(plainText), 'utf8'), cipher.final()]);
-    const tag = cipher.getAuthTag();
-    return 'enc:' + Buffer.from(iv).toString('base64') + ':' + Buffer.from(tag).toString('base64') + ':' + Buffer.from(ciphertext).toString('base64');
-  } catch (e) {
-    logger.error(e, 'Failed to encrypt secret; storing plaintext');
-    return plainText;
-  }
-}
-
-function decryptSecret(stored) {
-  if (stored == null) return null;
-  try {
-    if (typeof stored === 'string' && stored.startsWith('enc:')) {
-      if (!encryptionKey) {
-        logger.warn('CREDENTIALS_KEY not set; cannot decrypt stored password');
-        return null;
-      }
-      const [, ivB64, tagB64, dataB64] = stored.split(':');
-      const iv = Buffer.from(ivB64, 'base64');
-      const tag = Buffer.from(tagB64, 'base64');
-      const data = Buffer.from(dataB64, 'base64');
-      const decipher = createDecipheriv('aes-256-gcm', encryptionKey, iv);
-      decipher.setAuthTag(tag);
-      const plaintext = Buffer.concat([decipher.update(data), decipher.final()]).toString('utf8');
-      return plaintext;
-    }
-    return stored; // plaintext fallback
-  } catch (e) {
-    logger.error(e, 'Failed to decrypt secret');
-    return null;
-  }
-}
 
 class Metadata extends Model {
 
@@ -640,77 +597,7 @@ Instance.hasOne(PhoneNumber, { foreignKey: 'instanceId', onDelete: 'SET NULL', a
 
 
 // Registration-based endpoints (non-DDI)
-class PhoneRegistration extends Model {}
-
-PhoneRegistration.init({
-  id: {
-    type: DataTypes.UUID,
-    primaryKey: true,
-    defaultValue: DataTypes.UUIDV4
-  },
-  name: {
-    type: DataTypes.STRING,
-    allowNull: true
-  },
-  handler: {
-    type: DataTypes.STRING,
-    allowNull: false,
-    defaultValue: 'livekit'
-  },
-  outbound: {
-    type: DataTypes.BOOLEAN,
-    defaultValue: false
-  },
-  registrar: {
-    type: DataTypes.STRING,
-    allowNull: false
-  },
-  username: {
-    type: DataTypes.STRING,
-    allowNull: false
-  },
-  password: {
-    type: DataTypes.TEXT,
-    allowNull: false,
-    set(value) {
-      // Avoid double-encryption if it already looks encrypted
-      const toStore = (typeof value === 'string' && value.startsWith('enc:')) ? value : encryptSecret(value);
-      this.setDataValue('password', toStore);
-    },
-    get() {
-      const raw = this.getDataValue('password');
-      return decryptSecret(raw);
-    }
-  },
-  options: {
-    type: DataTypes.JSONB,
-    allowNull: true
-  },
-  status: {
-    type: DataTypes.ENUM,
-    values: ['active', 'failed', 'disabled'],
-    defaultValue: 'active'
-  },
-  state: {
-    type: DataTypes.ENUM,
-    values: ['initial', 'registering', 'registered', 'failed'],
-    defaultValue: 'initial'
-  },
-  error: {
-    type: DataTypes.TEXT,
-    allowNull: true
-  },
-  lastSeenAt: {
-    type: DataTypes.DATE,
-    allowNull: true
-  }
-}, {
-  sequelize,
-  timestamps: true,
-  underscored: true,
-  charset: 'utf8',
-  collate: 'utf8_general_ci',
-});
+const PhoneRegistration = initPhoneRegistration(sequelize, DataTypes);
 
 
 class User extends Model {
